add curly braces around eq in efficient-mendler
note: curly braces are used in other files to wrap the equality type
but for some reason are not used in Cast.ced (the only file where
this occurs). Moreover, this code does not check with the prerelease
version of Cedille anyway, whereas Cedille 1.0 and above require
the curly braces.